+++ mediation/src/test/java/net/pubnative/mediation/request/PubnativeNetworkVideoTest.java
@@ -210,20 +210,7 @@ public class PubnativeNetworkVideoTest {
         request.mHandler = new Handler();
         request.mListener = listener;
 
-        request.load(null, "app_token", "testPlacementName");
-
-        verify(listener).onPubnativeNetworkVideoLoadFail(eq(request), eq(PubnativeException.VIDEO_PARAMETERS_INVALID));
-    }
-
-    @Test
-    public void load_withNullAppToken_callBackLoadFail() {
-
-        PubnativeNetworkVideo request = spy(PubnativeNetworkVideo.class);
-        PubnativeNetworkVideo.Listener listener = spy(PubnativeNetworkVideo.Listener.class);
-        request.mHandler = new Handler();
-        request.mListener = listener;
-
-        request.load(RuntimeEnvironment.application.getApplicationContext(), null, "testPlacementName");
+        request.load(null, "testPlacementName");
 
         verify(listener).onPubnativeNetworkVideoLoadFail(eq(request), eq(PubnativeException.VIDEO_PARAMETERS_INVALID));
     }
@@ -236,20 +223,7 @@ public class PubnativeNetworkVideoTest {
         request.mHandler = new Handler();
         request.mListener = listener;
 
-        request.load(RuntimeEnvironment.application.getApplicationContext(), "app_token", null);
-
-        verify(listener).onPubnativeNetworkVideoLoadFail(eq(request), eq(PubnativeException.VIDEO_PARAMETERS_INVALID));
-    }
-
-    @Test
-    public void load_withEmptyAppToken_callBackLoadFail() {
-
-        PubnativeNetworkVideo request = spy(PubnativeNetworkVideo.class);
-        PubnativeNetworkVideo.Listener listener = spy(PubnativeNetworkVideo.Listener.class);
-        request.mHandler = new Handler();
-        request.mListener = listener;
-
-        request.load(RuntimeEnvironment.application.getApplicationContext(), "", "testPlacementName");
+        request.load(RuntimeEnvironment.application.getApplicationContext(), null);
 
         verify(listener).onPubnativeNetworkVideoLoadFail(eq(request), eq(PubnativeException.VIDEO_PARAMETERS_INVALID));
     }
@@ -262,37 +236,9 @@ public class PubnativeNetworkVideoTest {
         request.mHandler = new Handler();
         request.mListener = listener;
 
-        request.load(RuntimeEnvironment.application.getApplicationContext(), "app_token", "");
+        request.load(RuntimeEnvironment.application.getApplicationContext(), "");
 
         verify(listener).onPubnativeNetworkVideoLoadFail(eq(request), eq(PubnativeException.VIDEO_PARAMETERS_INVALID));
-    }
-
-    @Test
-    public void load_withValidListenerAndTrueIsLoading_callBackLoadFail() {
-
-        PubnativeNetworkVideo request = spy(PubnativeNetworkVideo.class);
-        PubnativeNetworkVideo.Listener listener = spy(PubnativeNetworkVideo.Listener.class);
-        request.mHandler = new Handler();
-        request.mListener = listener;
-        request.mIsLoading = true;
-
-        request.load(RuntimeEnvironment.application.getApplicationContext(), "app_token", "testPlacementName");
-
-        verify(listener).onPubnativeNetworkVideoLoadFail(eq(request), eq(PubnativeException.VIDEO_LOADING));
-    }
-
-    @Test
-    public void load_withValidListenerAndTrueIsShown_callBackLoadFail() {
-
-        PubnativeNetworkVideo request = spy(PubnativeNetworkVideo.class);
-        PubnativeNetworkVideo.Listener listener = spy(PubnativeNetworkVideo.Listener.class);
-        request.mHandler = new Handler();
-        request.mListener = listener;
-        request.mIsShown = true;
-
-        request.load(RuntimeEnvironment.application.getApplicationContext(), "app_token", "testPlacementName");
-
-        verify(listener).onPubnativeNetworkVideoLoadFail(eq(request), eq(PubnativeException.VIDEO_SHOWN));
     }
 
     @Test
@@ -301,16 +247,7 @@ public class PubnativeNetworkVideoTest {
         PubnativeNetworkVideo request = spy(PubnativeNetworkVideo.class);
         request.mHandler = new Handler();
 
-        request.load(null, "app_token", "testPlacementName");
-    }
-
-    @Test
-    public void load_withNullAppTokenNullListener_pass() {
-
-        PubnativeNetworkVideo request = spy(PubnativeNetworkVideo.class);
-        request.mHandler = new Handler();
-
-        request.load(RuntimeEnvironment.application.getApplicationContext(), null, "testPlacementName");
+        request.load(null, "testPlacementName");
     }
 
     @Test
@@ -319,7 +256,7 @@ public class PubnativeNetworkVideoTest {
         PubnativeNetworkVideo request = spy(PubnativeNetworkVideo.class);
         request.mHandler = new Handler();
 
-        request.load(RuntimeEnvironment.application.getApplicationContext(), "app_token", null);
+        request.load(RuntimeEnvironment.application.getApplicationContext(), null);
     }
 
     @Test
